Move `doc` directory to `document`
