+++ src/document/doc-utils.js
