+++ src/language-js/embed.js
@@ -14,7 +14,7 @@ const {
     dedentToRoot
   },
   utils: { mapDoc, stripTrailingHardline }
-} = require("../doc");
+} = require("../document");
 
 function embed(path, print, textToDoc, options) {
   const node = path.getValue();
