+++ src/language-markdown/embed.js
@@ -5,7 +5,7 @@ const support = require("../main/support");
 const {
   builders: { hardline, literalline, concat, markAsRoot },
   utils: { mapDoc }
-} = require("../doc");
+} = require("../document");
 const { getFencedCodeBlockValue } = require("./utils");
 
 function embed(path, print, textToDoc, options) {
