+++ src/standalone.js
@@ -6,7 +6,7 @@ const core = require("./main/core");
 const getSupportInfo = require("./main/support").getSupportInfo;
 const sharedUtil = require("./common/util-shared");
 
-const doc = require("./doc");
+const doc = require("./document");
 
 const internalPlugins = [
   require("./language-css"),
